fix(node-ui): extend sub-graph chip + disambiguated result copy to MemoryLayerView
After #706 surfaced both root + sub-graph assertions in the same WM
list, two rows literally labeled `draft` are indistinguishable in
the `MemoryLayerView` `AssertionList` row, even though the promote
and preview handlers already key on `graphUri` and route correctly.
The `› <slug>` chip closed the same gap for `AssertionsList`
(components.tsx) in the original #706; we missed the second
consumer.

Three small changes:

  1. Lift `truncateMiddle` from `components.tsx` into
     `lib/truncate.ts` so both consumers share one definition —
     same `lib/` pattern as `sub-graph-uri.ts` from the prior
     PR #710 fix.
  2. Render the chip in the `MemoryLayerView` row using the
     identical JSX pattern from `AssertionsList` (same class,
     same `›` glyph, same 18-char mid-truncation, same tooltip).
  3. Disambiguate the success message by carrying `subGraph` on
     `promoteResult` and rendering `" (in <slug>)"` after the
     assertion name when set. Root rows keep the prior wording.
